requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
lxml==4.9.3
Pillow==10.0.1
reportlab==4.0.4